Speed up builds, use shallow clone for cpp-driver.
For CI flows, reducing bandwidth and cpu consumption to
checkout/build dependencies is important, so change cpp-driver to use a
shallow clone so as to not download the entire history of the
repository.

A side affect of this is that the cpp-driver git reference must be
specified by branch/tag instead of commit sha.